Issue 14: Investigate: colour tests fail in Git Bash
Updated usage to make it clear that colors are not sent to pipes
and made it enable colors on windows when the option is specified.
So colors work in cmd and the tests pass. Colors don't work in
Git Bash and the color tests fail because Git Bash is doing
something shady.